extract User#can_follow? method, and not display message when make fellowship failed

+++ app/controllers/fellowships_controller.rb
@@ -6,32 +6,13 @@ class FellowshipsController < ApplicationController
 
   def create
     user = User.find(params[:user_id])
-
-    if user == current_user
-      flash[:danger] = "You can't follow yourself."
-      redirect_to people_path and return
-    end
-
-    fellowship = current_user.fellowships.build(followed_user: user)
-    if fellowship.save
-      flash[:success] = 'Follow success.'
-    else
-      flash[:info] = 'You are already following he/she.'
-    end
-
+    current_user.fellowships.create(followed_user: user) if current_user.can_follow?(user)
     redirect_to people_path
   end
 
   def destroy
     fellowship = current_user.fellowships.find_by(id: params[:id])
-
-    if fellowship.present?
-      fellowship.destroy
-      flash[:success] = 'Unfollow success.'
-    else
-      flash[:warning] = 'you can not do this!'
-    end
-
+    fellowship.destroy if fellowship.present?
     redirect_to people_path
   end
 end
